Rename AnnotationsSummary -> AnnotationsTable
